requirements.txt
PyYAML==3.12
dnspython==1.15.0
sleekxmpp==1.3.3
schedule==0.5.0
